pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "pyflic"
dynamic = ["version"]
description = "FLIC (Fly Liquid-food Interaction Counter) data analysis toolkit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "plotnine>=0.15.3",
    "pyqt6>=6.10.2",
    "pyyaml>=6.0.3",
    "statsmodels>=0.14.6",
]

[project.scripts]
pyflic = "pyflic.__main__:main"
pyflic-config = "pyflic.base.config_editor:launch"
pyflic-qc = "pyflic.base.qc_viewer:main"
pyflic-hub = "pyflic.base.analysis_hub:main"
pyflic-lint = "pyflic.base.yaml_lint:main_cli"

[tool.hatch.version]
source = "vcs"

[dependency-groups]
dev = [
    "hypothesis>=6.111.0",
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
]
